Add files via upload
I know this is a JavaScript Repository BUT I know nothing about JavaScript and this project really helped teach me that I can't run before I learn to walk. I will eventually return and remake or finish this project but for now I'll leave it on unfinished, I'm not quitting I'm acknowledging that I need to study ,practice ,learn and educate myself before I try projects of this difficulty. (Or else it's because I was following the tutorial by hearing alone)